requirements.txt
asgiref==3.2.3
django-environ==0.4.5
django==3.0.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.1
requests==2.22.0
django-filter==2.2.0
django-admin-interface==0.13.5
rollbar==0.15.0
pandas==1.1.5
xlrd==1.2.0
pyreadstat==1.0.8
xlsxwriter
